+++ components/VideosRow/VideosRow.stories.tsx
@@ -83,6 +83,28 @@ Open.args = {
       published: new Date(),
       title: 'Boris Johnson Visits the Godfather: Parody',
       fullSlug: ''
+    },
+    {
+      id: 10,
+      thumbnail_image: {
+        url:
+          'https://res.cloudinary.com/split-the-facts/image/upload/v1626856861/Screenshot_2021_07_18_at_08_46_02_77df809cff.png'
+      },
+      embed_url: 'tq1G9Z66Nds',
+      published: new Date(),
+      title: 'Boris Johnson Visits the Godfather: Parody',
+      fullSlug: ''
+    },
+    {
+      id: 15,
+      thumbnail_image: {
+        url:
+          'https://res.cloudinary.com/split-the-facts/image/upload/v1626856861/Screenshot_2021_07_18_at_08_46_02_77df809cff.png'
+      },
+      embed_url: 'tq1G9Z66Nds',
+      published: new Date(),
+      title: 'Boris Johnson Visits the Godfather: Parody',
+      fullSlug: ''
     }
   ]
 }