+++ next-seo.config.js
@@ -1,4 +1,4 @@
 export default {
-  title: 'SPlit the facts!!!',
+  title: 'Split the facts!',
   description: 'Parody comedic sketches for your entertainment. Enjoy'
 }